% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytes.R
\name{as_bytes}
\alias{as_bytes}
\title{Coerce an object to a bytes object}
\usage{
as_bytes(x)
}
\arguments{
\item{x}{Object to be coerced}
}
\description{
Coerce an object to a bytes object
}
\examples{
as_bytes("1KB") < "1MB"
}
