% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_bytes.R
\name{as_fs_bytes}
\alias{as_fs_bytes}
\title{Coerce an object to a fs_bytes object}
\usage{
as_fs_bytes(x)
}
\arguments{
\item{x}{Object to be coerced}
}
\description{
Coerce an object to a fs_bytes object
}
\examples{
as_fs_bytes("1KB") < "1MB"
}
