% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_with_deduplication}
\alias{call_with_deduplication}
\title{Calls a function via deduplication of the input rather than each individual element.}
\usage{
call_with_deduplication(func, x, ...)
}
\arguments{
\item{func}{Function to call.}

\item{x}{First unnamed argument to the function, and the one that should be deduplicated.}

\item{...}{Other arguments to the function.}
}
\value{
\code{func} applied to \code{x}.
}
\description{
Used to speed up some functions.
}
