% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_copy}
\alias{dir_copy}
\title{Copy a directory}
\usage{
dir_copy(path, new_path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{new_path}{new location to copy to.}
}
\description{
Copy a directory
}
