% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_create}
\alias{dir_create}
\title{Create a directory}
\usage{
dir_create(path, mode = "u+rwx,go+rx")
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{mode}{If file is created, what mode should it have?}
}
\description{
Create a directory
}
\examples{
x <- tempfile()
try(is_dir(x))
dir_create(x)
is_dir(x)
}
