% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_delete}
\alias{dir_delete}
\title{Delete files in a directory}
\usage{
dir_delete(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\description{
Delete files in a directory
}
