% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_exists}
\alias{dir_exists}
\title{Check if a directory exists}
\usage{
dir_exists(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\description{
Check if a directory exists
}
