% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_info}
\alias{dir_info}
\title{Directory information}
\usage{
dir_info(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\value{
A tibble with metadata for each file
}
\description{
Directory information
}
