% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_list.R
\name{dir_list}
\alias{dir_list}
\alias{dir_walk}
\title{List files in a directory}
\usage{
dir_list(path = ".", all = FALSE, recursive = TRUE, type = "any",
  pattern = NULL, ...)

dir_walk(path = ".", f, all = FALSE, recursive = TRUE, type = "any",
  pattern = NULL, ...)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{all}{If \code{TRUE} hidden files are also returned.}

\item{recursive}{Should directories be listed recursively?}

\item{type}{File type to return, one of "any", "file", "directory",
"symlink", "FIFO", "socket", "character_device" or "block_device".}

\item{pattern}{A regular expression pattern used passed to \link{grep} to filter
the filenames.}

\item{...}{Additional arguments passed to \link{grep}.}

\item{f}{A function, taking one parameter, the current path entry.}
}
\description{
List files in a directory
}
\section{Functions}{
\itemize{
\item \code{dir_walk}: Walk along a directory, calling \code{f} for each entry in
the directory.
}}

\examples{
dir_list(system.file())
dir_walk(system.file(), function(p) if (grepl("profile", p)) print(p))
}
