% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_access}
\alias{file_access}
\title{Query files for access permissions}
\usage{
file_access(path, mode = "exists")
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{mode}{A character vector containing one or more of 'exists', 'read',
'write', 'execute'.}
}
\value{
A logical vector
}
\description{
Query files for access permissions
}
