% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_chown}
\alias{file_chown}
\title{Change ownership or group of a file}
\usage{
file_chown(path, user_id = NULL, group_id = NULL)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{user_id}{The user id of the new owner, the R process must be privlaged to change
this.}

\item{group_id}{The group id of the new owner}
}
\description{
Change ownership or group of a file
}
