% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_create.R
\name{file_create}
\alias{file_create}
\title{Create a file}
\usage{
file_create(path, mode = "u+rw,go+r")
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{mode}{If file is created, what mode should it have?}
}
\description{
Unlike \code{\link[=file.create]{file.create()}}, will not truncate an existing file.
}
\examples{
x <- file_create(tempfile())
x
}
