% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_delete}
\alias{file_delete}
\title{Delete a file
template fs}
\usage{
file_delete(path)
}
\description{
Delete a file
template fs
}
\examples{
x <- file_create(tempfile())
file_exsts(x)
file_delete(x)
file_exists(x)
}
