% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_exists}
\alias{file_exists}
\title{Check if a file exists}
\usage{
file_exists(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\description{
Check if a file exists
}
