% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_info}
\alias{file_info}
\title{Query file metadata}
\usage{
file_info(path)
}
\arguments{
\item{path}{A character vector of paths.}
}
\value{
A data.frame with metadata for each file
}
\description{
Query file metadata
}
\examples{
write.csv(mtcars, "mtcars.csv")
file_info("mtcars.csv")
}
