% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_move.R
\name{file_move}
\alias{file_move}
\title{Move or rename files}
\usage{
file_move(path, new_path)
}
\arguments{
\item{path}{A character vector of paths.}

\item{new_path}{New file path. If \code{new_path} is existing directory, the file
will be moved into that directory; otherwise it will be moved/renamed to
the full path.

Should either be the same length as \code{path}, or a single directory.}
}
\description{
Move or rename files
}
