% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_show}
\alias{file_show}
\title{Open the file(s) directory in an interactive explorer}
\usage{
file_show(path = ".")
}
\arguments{
\item{path}{A character vector of paths.}
}
\value{
The directories that were opened (invisibly).
}
\description{
Open the file(s) directory in an interactive explorer
}
