% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmode.R
\name{as_fmode}
\alias{as_fmode}
\alias{fmode}
\title{File permissions}
\usage{
as_fmode(x)

fmode(x)
}
\arguments{
\item{x}{An object which is to be coerced to a fmode object. Can be an
number or hexidecimal character representation, including symbolic
representations.}
}
\description{
File permissions
}
