% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-package.R
\docType{package}
\name{fs-package}
\alias{fs}
\alias{fs-package}
\title{fs: Cross-Platform File System Operations Based on 'libuv'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A cross-platform interface to file system operations, built on top of the 'libuv' C library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fs.r-lib.org}
  \item \url{https://github.com/r-lib/fs}
  \item Report bugs at \url{https://github.com/r-lib/fs/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Jim Hester
  \item Hadley Wickham \email{hadley@posit.co}
}

Other contributors:
\itemize{
  \item libuv project contributors (libuv library) [copyright holder]
  \item Joyent, Inc. and other Node contributors (libuv library) [copyright holder]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
