% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_perms.R
\name{fs_perms}
\alias{fs_perms}
\alias{as_fs_perms}
\alias{fs_perms}
\title{Create, modify and view file permissions}
\usage{
as_fs_perms(x)

fs_perms(x)
}
\arguments{
\item{x}{An object which is to be coerced to a fs_perms object. Can be an
number or octal character representation, including symbolic
representations.}
}
\description{
\code{fs_perms()} objects help one create and modify file permissions easily.
They support both numeric input, octal and symbolic character
representations. Compared to \link{octmode} they support symbolic representations
and display the mode the same format as \code{ls} on POSIX systems.
}
\examples{
# Integer and numeric
fs_perms(420L)
fs_perms(c(511, 420))

# Octal
fs_perms("777")
fs_perms(c("777", "644"))

# Symbolic
fs_perms("a+rwx")
fs_perms(c("a+rwx", "u+rw,go+r"))
}
