% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realpath.R
\name{fs_realpath}
\alias{fs_realpath}
\alias{dir_realpath}
\title{Return the canonicalized absolute pathname}
\usage{
fs_realpath(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\value{
\code{[character(1)]} the fully resolved path.
}
\description{
This is functionally equivalent to \code{\link[base:normalizePath]{base::normalizePath()}}.
}
