% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{id}
\alias{id}
\alias{group_ids}
\alias{user_ids}
\title{Lookup Users and Groups on a system}
\usage{
group_ids()

user_ids()
}
\description{
These functions use the GETPWENT(3) and GETGRENT(3) system calls to query
users and groups respectively. They return their results in a \code{data.frame}.
}
