% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_absolute_path}
\alias{is_absolute_path}
\title{Test if a path is an absolute path}
\usage{
is_absolute_path(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
Test if a path is an absolute path
}
\examples{
is_absolute_path("/foo")
is_absolute_path("C:\\\\foo")
is_absolute_path("\\\\\\\\myserver\\\\foo\\\\bar")

is_absolute_path("foo/bar")
}
