% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_dir_empty}
\alias{is_dir_empty}
\title{Check if a directory is empty}
\usage{
is_dir_empty(path)
}
\arguments{
\item{path}{A character string specifying the path to the directory to check.}
}
\value{
A logical value. Returns \code{TRUE} if the directory is empty,
\code{FALSE} otherwise.
}
\description{
This function checks whether a given directory is empty or not.
}
\examples{
\dontrun{
is_dir_empty("path/to/empty/directory")  # Returns TRUE
is_dir_empty("path/to/non-empty/directory")  # Returns FALSE
}

}
