% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link_create}
\alias{link_create}
\title{Create a file link}
\usage{
link_create(path, new_path, symbolic = TRUE)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{new_path}{Character vector of locations the link should point to.}

\item{symbolic}{Boolean value determining if the link should be a symbolic
(the default) or hard link.}
}
\description{
Create a file link
}
