% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link_info}
\alias{link_info}
\title{Link information}
\usage{
link_info(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\value{
A tibble with metadata for each file
}
\description{
Link information
}
