% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link_path}
\alias{link_path}
\title{Read the value of a symbolic link}
\usage{
link_path(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\description{
Read the value of a symbolic link
}
\examples{
file_create("foo")
link_create("foo", "bar")
link_path("bar")
file_delete(c("foo", "bar"))
}
