% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_common}
\alias{path_common}
\title{Find the common parts of two (or more) paths}
\usage{
path_common(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
The input paths must be either all relative or all absolute.
}
