% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_ext}
\alias{path_ext}
\alias{path_ext_remove}
\alias{path_ext_set}
\alias{path_ext<-}
\title{Manipulate path extensions}
\usage{
path_ext(path)

path_ext_remove(path)

path_ext_set(path, ext)

path_ext(path) <- value
}
\arguments{
\item{path}{A character vector of paths.}

\item{ext, value}{The new file extension.}
}
\description{
\code{path_ext()} returns the last extension (if any) for a path.
\code{path_ext_remove()} removes the last extension and returns the rest of the
path. \code{path_ext_set()} replaces the extension with a new extension. If there
is no existing extension the new extension is appended.
}
\examples{
path_ext("file.zip")
path_ext("file.tar.gz")

path_ext_remove("file.tar.gz")

# Only one level of extensions is removed
path_ext_set(path_ext_remove("file.tar.gz"), "zip")
}
