% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_home}
\alias{path_home}
\title{Provide the path to the users home directory}
\usage{
path_home(...)
}
\arguments{
\item{...}{Additional paths appended to the temporary directory by \code{path()}.}
}
\description{
Provide the path to the users home directory
}
