% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_norm}
\alias{path_norm}
\title{Normalize a path}
\usage{
path_norm(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
\code{path_norm()} collapses redundant separators and up-level references, so \code{A//B},
\code{A/B}, \code{A/.B} and \code{A/foo/../B} all become \code{A/B}. If one of the paths is a
symbolic link, this may change the meaning of the path, in this case one can
use \code{path_realize()} beforehand to follow the symlink.
}
