% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_realize}
\alias{path_realize}
\title{Return the canonicalized absolute pathname}
\usage{
path_realize(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\value{
\code{[character(1)]} the fully resolved path.
}
\description{
This is functionally equivalent to \code{\link[base:normalizePath]{base::normalizePath()}}.
}
