% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_temp}
\alias{path_temp}
\title{Path to sessions temporary directory}
\usage{
path_temp(...)
}
\arguments{
\item{...}{Additional paths appended to the temporary directory by \code{path()}.}
}
\description{
Analogous to \code{\link[base:tempdir]{base::tempdir()}}.
}
