% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_tidy}
\alias{path_tidy}
\title{Tidy paths}
\usage{
path_tidy(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\value{
An \code{fs_path} object, which is a character vector that also has class
\code{fs_path}
}
\description{
untidy paths are all different, tidy paths are all the same.
Tidy paths always use \code{/} to delimit directories, never have
multiple \code{/} or trailing \code{/} and have colourised output based on the file
type.
}
