% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_ls}
\alias{zip_ls}
\title{List zip contents}
\usage{
zip_ls(path, method = "internal")
}
\arguments{
\item{path}{A character vector of zip archives.}

\item{method}{The method to be used. An alternative is to use
\code{getOption("unzip")}, which on a Unix-alike may be set to the path to a
unzip program.}
}
\description{
List zip contents
}
