% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_move}
\alias{zip_move}
\title{Extract zip contents}
\usage{
zip_move(
  zip,
  dir = NULL,
  files = NULL,
  overwrite = TRUE,
  junk = FALSE,
  method = "internal"
)
}
\arguments{
\item{zip}{A character vector of zip archives.}

\item{dir}{The directory to extract files to (the equivalent of \code{unzip -d}).
It will be created if necessary.}

\item{files}{A character vector of paths to be extracted: the default is to
extract all files.}

\item{overwrite}{If \code{TRUE} (default), overwrite existing files (the
equivalent of \code{unzip -o}), otherwise ignore such files (the equivalent of
\code{unzip -n}).}

\item{junk}{If \code{TRUE}, use only the basename of the stored path when
extracting. The equivalent of \code{unzip -j}.}

\item{method}{The method to be used. An alternative is to use
\code{getOption("unzip")}, which on a Unix-alike may be set to the path to a
unzip program.}
}
\value{
The path to the extracted files (invisibly).
}
\description{
Extract zip contents
}
