% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\alias{file_delete}
\alias{dir_delete}
\alias{link_delete}
\title{Delete a file, directory, or link.}
\usage{
file_delete(path)

dir_delete(path)

link_delete(path)
}
\arguments{
\item{path}{A character vector of paths.}
}
\value{
The deleted paths.
}
\description{
\code{dir_delete()} will first delete the contents of the directory, then remove
the directory.
}
\examples{
\dontshow{fs:::pkgdown_tmp("/tmp/filedd4635bd7c86")}
x <- file_create(file_temp())
file_exists(x)
file_delete(x)
file_exists(x)
}
