% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R, R/dir.R
\name{dir_create}
\alias{dir_create}
\alias{file_create}
\title{Create a file or directory.}
\usage{
dir_create(path, mode = "u+rwx,go+rx")

file_create(path, mode = "u+rw,go+r")

dir_create(path, mode = "u+rwx,go+rx")
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{mode}{If file/directory is created, what mode should it have?}
}
\description{
These functions ensure that \code{path} exists; if it already exists it will
be left unchanged. That means that ompared to \code{\link[=file.create]{file.create()}},
\code{file_create()} will not truncate an existing, and compared to
\code{\link[=dir.create]{dir.create()}}, \code{dir_create()} will silently ignore
existing directories.
}
\examples{
x <- file_create(tempfile())
is_file(x)
# dir_create applied to the same path will fail
try(dir_create(x))

x <- dir_create(tempfile())
is_dir(x)
# file_create applied to the same path will fail
try(file_create(x))
x <- tempfile()
try(is_dir(x))
dir_create(x)
is_dir(x)
}
