% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_list}
\alias{dir_list}
\title{List files in a directory}
\usage{
dir_list(path = ".", recursive = TRUE, type = "any", pattern = NULL,
  ...)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{recursive}{Should directories be listed recursively?}

\item{type}{File type to return, one of "any", "file", "directory",
"symlink", "FIFO", "socket", "character_device" or "block_device".}

\item{pattern}{A regular expression pattern used passed to \link{grep} to filter
the filenames.}

\item{...}{Additional arguments passed to \link{grep}.}
}
\description{
List files in a directory
}
\examples{
dir_list(system.file())
}
