% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_move}
\alias{dir_move}
\title{Move a directory}
\usage{
dir_move(path, new_path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{new_path}{new file path}
}
\description{
Move a directory
}
