% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{dir_tree}
\alias{dir_tree}
\title{Print contents of directories in a tree-like format}
\usage{
dir_tree(path = ".", recurse = TRUE, ...)
}
\arguments{
\item{path}{A path to print the tree from}

\item{recurse}{If \code{TRUE} recurse fully, if a positive number the number of levels
to recurse.}

\item{...}{Additional arguments passed to \link{dir_ls}.}
}
\description{
Print contents of directories in a tree-like format
}
