% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{dir_tree}
\alias{dir_tree}
\title{Print contents of directories in a tree-like format}
\usage{
dir_tree(path = ".", recursive = TRUE, ...)
}
\arguments{
\item{path}{A path to print the tree from}

\item{recursive}{Should directories be listed recursively?
the filenames.}

\item{...}{Additional arguments passed to \link{dir_ls}.}
}
\description{
Print contents of directories in a tree-like format
}
