% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_chmod}
\alias{file_chmod}
\title{Change file permissions}
\usage{
file_chmod(path, mode)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{mode}{A character representation of the mode, in either hexidecimal or symbolic format.}
}
\description{
Change file permissions
}
\details{
\strong{Cross-compatibility warning:} File permissions are different on
Windows, so many of the examples below may not behave as you expect. First,
there is no executable bit on Windows, so any attempt to change this will
have no affect. Second, you can only modify user permissions (\code{u}).
Attempting to modify the permissions for group (\code{g}) or others (\code{o}) will
throw an error. Modifying the permissions of all users (\code{a}) will only
affect the user permissions (\code{u}). If you have installed a bash emulator on
your Windows machine, e.g. \href{https://gitforwindows.org/}{Git Bash}, you can
confirm this behavior by running \code{chmod} and observing that it's not
possible to modify the permissions for group (\code{g}) or others (\code{o}).
}
\examples{
\dontshow{.old_wd <- setwd(tempdir())}
file_create("foo", mode = "000")
file_chmod("foo", "777")
file_info("foo")$permissions

file_chmod("foo", "u-x")
file_info("foo")$permissions

file_chmod("foo", "a-wrx")
file_info("foo")$permissions

file_chmod("foo", "u+wr")
file_info("foo")$permissions

# It is also vectorized
files <- c("foo", file_create("bar", mode = "000"))
file_chmod(files, "a+rwx")
file_info(files)$permissions

file_chmod(files, c("644", "600"))
file_info(files)$permissions
\dontshow{setwd(.old_wd)}
}
