% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_chmod}
\alias{file_chmod}
\title{Change file permissions}
\usage{
file_chmod(path, mode)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{mode}{A character representation of the mode, in either hexidecimal or symbolic format.}
}
\description{
Change file permissions
}
\examples{
x <- file_create(tempfile(), "000")
file_chmod(x, "777")
file_info(x)$permissions

file_chmod(x, "u-x")
file_info(x)$permissions

file_chmod(x, "a-wrx")
file_info(x)$permissions

file_chmod(x, "ug+wr")
file_info(x)$permissions
}
