% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_chown}
\alias{file_chown}
\title{Change ownership or group of a file}
\usage{
file_chown(path, user_id = NULL, group_id = NULL)
}
\arguments{
\item{path}{A character vector of paths.}

\item{user_id}{The user id of the new owner, specified as a numeric ID or
name. The R process must be privileged to change this.}

\item{group_id}{The group id of the new owner, specified as a numeric ID or
name.}
}
\description{
Change ownership or group of a file
}
