% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_copy}
\alias{file_copy}
\title{Copy a file}
\usage{
file_copy(path, new_path, overwrite = FALSE)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{new_path}{Character vector of paths to the new files.}

\item{overwrite}{Overwrite files if they exist. If this is \code{FALSE} and a
file exists and error will be thrown.}
}
\description{
Copy a file
}
\examples{
file_create("foo")
file_copy("foo", "bar")
try(file_copy("foo", "bar"))
file_copy("foo", "bar", force = TRUE)
file_delete(c("foo", "bar"))
}
