% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{file_create}
\alias{file_create}
\alias{dir_create}
\alias{link_create}
\title{Create a file, directory, or link.}
\usage{
file_create(path, mode = "u+rw,go+r")

dir_create(path, mode = "u+rwx,go+rx", recursive = TRUE)

link_create(path, new_path, symbolic = TRUE)
}
\arguments{
\item{path}{A character vector of paths.}

\item{mode}{If file/directory is created, what mode should it have?
Links do not have permissions; they inherit the permissions of the
file they link to.}

\item{recursive}{should intermediate directories be created if they do not
exist?}

\item{symbolic}{Boolean value determining if the link should be a symbolic
(the default) or hard link.}
}
\description{
These functions ensure that \code{path} exists; if it already exists it will
be left unchanged. That means that ompared to \code{\link[=file.create]{file.create()}},
\code{file_create()} will not truncate an existing, and compared to
\code{\link[=dir.create]{dir.create()}}, \code{dir_create()} will silently ignore
existing directories.
}
\examples{
x <- file_create(tempfile())
is_file(x)
# dir_create applied to the same path will fail
try(dir_create(x))

x <- dir_create(tempfile())
is_dir(x)
# file_create applied to the same path will fail
try(file_create(x))
}
