% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_delete}
\alias{file_delete}
\title{Delete a file}
\usage{
file_delete(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\description{
Delete a file
}
\examples{
x <- file_create(tempfile())
file_exists(x)
file_delete(x)
file_exists(x)
}
