% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_info}
\alias{file_info}
\title{Query file metadata}
\usage{
file_info(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\value{
A tibble with metadata for each file
}
\description{
Query file metadata
}
