% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_move.R
\name{file_move}
\alias{file_move}
\title{Move a file}
\usage{
file_move(path, new_path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{new_path}{new file path}
}
\description{
Move a file
}
