% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_move}
\alias{file_move}
\title{Rename a file}
\usage{
file_move(path, new_path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}

\item{new_path}{new file path}
}
\value{
\code{[character()]}\cr UTF-8 encoded path.
}
\description{
Rename a file
}
