% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_show}
\alias{file_show}
\title{Open the file(s) directory in an interactive explorer}
\usage{
file_show(path = ".", browser = getOption("browser"))
}
\arguments{
\item{path}{A character vector of paths.}

\item{browser}{a non-empty character string giving the name of the
    program to be used as the HTML browser.  It should be in the PATH,
    or a full path specified.  Alternatively, an \R function to be
    called to invoke the browser.

    Under Windows \code{NULL} is also allowed (and is the default), and
    implies that the file association mechanism will be used.
  }
}
\value{
The directories that were opened (invisibly).
}
\description{
Open the file(s) directory in an interactive explorer
}
