% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_perms.R
\name{fs_perms}
\alias{fs_perms}
\alias{as_fmode}
\alias{fs_perms}
\title{File permissions}
\usage{
as_fmode(x)

fs_perms(x)
}
\arguments{
\item{x}{An object which is to be coerced to a fs_perms object. Can be an
number or octal character representation, including symbolic
representations.}
}
\description{
File permissions
}
\examples{
# Integer and numeric
fs_perms(420L)
fs_perms(c(511, 420))

# Octal
fs_perms("777")
fs_perms(c("777", "644"))

# Symbolic
fs_perms("a+rwx")
fs_perms(c("a+rwx", "u+rw,go+r"))
}
