% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmode.R
\name{fmode}
\alias{fmode}
\alias{as_fmode}
\alias{fmode}
\title{File permissions}
\usage{
as_fmode(x)

fmode(x)
}
\arguments{
\item{x}{An object which is to be coerced to a fmode object. Can be an
number or octal character representation, including symbolic
representations.}
}
\description{
File permissions
}
\examples{
# Integer and numeric
fmode(420L)
fmode(c(511, 420))

# Octal
fmode("777")
fmode(c("777", "644"))

# Symbolic
fmode("a+rwx")
fmode(c("a+rwx", "u+rw,go+r"))
}
