% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-package.R
\docType{package}
\name{fs-package}
\alias{fs}
\alias{fs-package}
\title{fs: Cross-Platform File System Operations Based on 'libuv'}
\description{
\ifelse{html}{
\out{
<a href="https://lifecycle.r-lib.org/articles/stages.html#maturing">
  <img src="https://img.shields.io/badge/lifecycle-maturing-blue.svg" alt="Maturing lifecycle">
</a>
}
}{
\url{https://lifecycle.r-lib.org/articles/stages.html#maturing}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fs.r-lib.org}
  \item \url{https://github.com/r-lib/fs}
  \item Report bugs at \url{https://github.com/r-lib/fs/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Jim Hester
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item libuv project contributors (libuv library) [copyright holder]
  \item Joyent, Inc. and other Node contributors (libuv library) [copyright holder]
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
