% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-package.R
\docType{package}
\name{fs-package}
\alias{fs}
\alias{fs-package}
\title{fs: Cross-Platform File System Operations Based on 'libuv'}
\description{
\ifelse{html}{
\out{
<a href="https://www.tidyverse.org/lifecycle/#stable">
  <img src="https://img.shields.io/badge/lifecycle-stable-green.svg" alt="Stable lifecycle">
</a>
}
}{
\url{https://www.tidyverse.org/lifecycle/#stable}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://fs.r-lib.org}
  \item \url{https://github.com/r-lib/fs}
  \item Report bugs at \url{https://github.com/r-lib/fs/issues}
}

}
\author{
\strong{Maintainer}: Jim Hester \email{james.f.hester@gmail.com}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item libuv project contributors (libuv library) [copyright holder]
  \item Joyent, Inc. and other Node contributors (libuv library) [copyright holder]
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
