% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_file}
\alias{is_file}
\alias{is_dir}
\alias{is_link}
\title{Functions to test for file types}
\usage{
is_file(path)

is_dir(path)

is_link(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\value{
A logical vector
}
\description{
Functions to test for file types
}
