% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link_path}
\alias{link_path}
\title{Read the value of a symbolic link}
\usage{
link_path(path)
}
\arguments{
\item{path}{A character vector of paths.}
}
\value{
A tidy path to the object the link points to.
}
\description{
Read the value of a symbolic link
}
\examples{
file_create("foo")
link_create(path_norm("foo"), "bar")
link_path("bar")
file_delete(c("foo", "bar"))
}
