% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\alias{path_wd}
\title{Construct path to a file or directory}
\usage{
path(..., ext = "")

path_wd(..., ext = "")
}
\arguments{
\item{...}{character vectors, if any values are NA, the result will also be
NA.}

\item{ext}{An optional extension to append to the generated path.}

\item{path}{A character vector of one or more paths.}
}
\description{
\code{path()} constructs a relative path, \code{path_wd()} constructs an absolute path
from the current working directory.
}
\examples{
path("foo", "bar", "baz", ext = "zip")

path("foo", letters[1:3], ext = "txt")
}
\seealso{
\code{\link[=path_home]{path_home()}}, \code{\link[=path_package]{path_package()}} for functions to construct paths
relative to the home and package directories respectively.
}
