% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Construct path to a file or directory}
\usage{
path(...)
}
\arguments{
\item{...}{character vectors}

\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\description{
Construct path to a file or directory
}
\seealso{
\code{\link[base:file.path]{base::file.path()}}
}
