% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Construct path to a file or directory}
\usage{
path(..., ext = "")
}
\arguments{
\item{...}{character vectors, if any values are NA, the result will also be
NA.}

\item{ext}{An optional extension to append to the generated path.}

\item{path}{A character vector of paths.}
}
\description{
Construct path to a file or directory
}
\seealso{
\code{\link[base:file.path]{base::file.path()}}
}
