% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_expand}
\alias{path_expand}
\title{Perform tilde expansion of a pathname}
\usage{
path_expand(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\value{
\code{[character()]}\cr UTF-8 encoded path.
}
\description{
equivalent to \link[base:path.expand]{base::path.expand}
}
