% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_expand}
\alias{path_expand}
\title{Perform tilde expansion of a pathname}
\usage{
path_expand(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
equivalent to \link[base:path.expand]{base::path.expand}
}
