% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_home}
\alias{path_home}
\alias{path_temp}
\title{Paths starting from useful directories}
\usage{
path_home(...)

path_temp(...)
}
\arguments{
\item{...}{Additional paths appended to the temporary directory by \code{path()}.}
}
\description{
\itemize{
\item \code{path_temp()} starts the path with the session temporary directory
\item \code{path_home()} starts the path with the expanded users home directory
}
}
\details{
On Windows the home directory is set as follows:
If environment variable \code{R_USER} is set, its value is used. Otherwise if
environment variable \code{HOME} is set, its value is used. After those two
user-controllable settings, R tries to find system-defined home directories.
It first tries to use the Windows "personal" directory (typically
\code{C:\\Users\\username\\Documents}). If that fails, if both environment
variables \code{HOMEDRIVE} and \code{HOMEPATH} are set (and they normally are), the
value is \code{${HOMEDRIVE}${HOMEPATH}}. If all of these fail, the current
working directory is used. Source: \href{https://cran.r-project.org/bin/windows/base/rw-FAQ.html#What-are-HOME-and-working-directories_003f}{R for Windows FAQ - 2.14}.
}
\examples{
path_home()
path_home("R")

path_temp()
path_temp("does-not-exist")
}
