% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_math}
\alias{path_math}
\alias{path_realize}
\alias{path_expand}
\alias{path_split}
\alias{path_join}
\alias{path_absolute}
\alias{path_norm}
\alias{path_relative}
\alias{path_common}
\title{Path computations}
\usage{
path_realize(path)

path_expand(path)

path_split(path)

path_join(parts)

path_absolute(path)

path_norm(path)

path_relative(path, start = ".")

path_common(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{parts}{A list of character vectors, corresponding to split paths.}

\item{start}{A starting directory to compute relative path to.}
}
\description{
Path computations
}
\section{Functions}{
\itemize{
\item \code{path_realize}: returns the canonical path, eliminating any symbolic
links.

\item \code{path_expand}: performs tilde expansion on a path, replacing instances of
\code{~} or \code{~user} with the user's home directory.

\item \code{path_split}: splits paths into parts.

\item \code{path_join}: joins parts together.

\item \code{path_absolute}: returns a normalized, absolute version of a path.

\item \code{path_norm}: collapses redundant separators and
up-level references, so \code{A//B}, \code{A/B}, \code{A/.B} and \code{A/foo/../B} all become
\code{A/B}. If one of the paths is a symbolic link, this may change the meaning
of the path, in this case one can use \code{path_realize()} beforehand to follow
the symlink.

\item \code{path_common}: Find the common parts of two (or more) paths
}}

