% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_norm}
\alias{path_norm}
\alias{dir_realpath}
\title{Return the canonicalized absolute pathname}
\usage{
path_norm(path)
}
\arguments{
\item{path}{A character vector of paths.}
}
\value{
\code{[character(1)]} the fully resolved path.
}
\description{
This is functionally equivalent to \code{\link[base:normalizePath]{base::normalizePath()}}.
}
