% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_split}
\alias{path_split}
\title{Split a path into components}
\usage{
path_split(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\value{
A list of separated paths
}
\description{
Split a path into components
}
