% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_split}
\alias{path_split}
\title{Split a path into components}
\usage{
path_split(path)
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\value{
\code{[character()]}\cr UTF-8 encoded path.
}
\description{
Split a path into components
}
