% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_temp}
\alias{path_temp}
\title{Path to sessions temporary directory}
\usage{
path_temp()
}
\arguments{
\item{path}{\code{[character(1)]}\cr Pathname, this will be UTF-8 encoded before passing to libuv.}
}
\value{
\code{[character()]}\cr UTF-8 encoded path.
}
\description{
Analogous to \code{\link[base:tempdir]{base::tempdir()}}.
}
