% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_create}
\alias{zip_create}
\title{Create zip archive}
\usage{
zip_create(
  path,
  zip = NULL,
  quiet = TRUE,
  junk = FALSE,
  method = Sys.getenv("R_ZIPCMD", "zip")
)
}
\arguments{
\item{path}{A character vector of files to archive.}

\item{zip}{The path of the zip file to create. If none is supplied, and only
one path is supplied, an archive with that name will be created.}

\item{quiet}{If \code{TRUE}, suppress deflating messages. The equivalent of
\code{zip -q}.}

\item{junk}{If \code{TRUE}, use only the basename of the path when
creating the archive. The equivalent of \code{zip -j}.}

\item{method}{A character string specifying the external command to be used.}
}
\value{
The path to the created archive (invisibly).
}
\description{
Create zip archive
}
