% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_create}
\alias{zip_create}
\title{Create zip archive}
\usage{
zip_create(path, zip = NULL, method = Sys.getenv("R_ZIPCMD", "zip"), ...)
}
\arguments{
\item{path}{A character vector of files to archive.}

\item{zip}{The path of the zip file to create. If none is supplied, and only
one path is supplied, an archive with that name will be created.}

\item{method}{A character string specifying the external command to be used.}

\item{...}{Additional arguments passed to \code{\link[utils:zip]{utils::zip()}}}
}
\description{
Create zip archive
}
