% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_create}
\alias{zip_create}
\title{Create zip archive}
\usage{
zip_create(
  path,
  zip = NULL,
  quiet = TRUE,
  junk = FALSE,
  extra = "",
  method = Sys.getenv("R_ZIPCMD", "zip")
)
}
\arguments{
\item{path}{A character vector of files to archive.}

\item{zip}{The path of the zip file to create. If none is supplied, and only
one path is, an archive with that \code{\link[=basename]{basename()}} will be created in the same
directory.}

\item{quiet}{If \code{TRUE}, suppress deflating messages (the equivalent of
\code{zip -q}).}

\item{junk}{If \code{TRUE}, use only the basename of the path when creating the
archive (the equivalent of \code{zip -j}).}

\item{extra}{An optional character vector passed to \code{\link[=system]{system()}} as arguments
of the external \code{zip} command (e.g., \code{-x} followed by paths to exclude).}

\item{method}{A character string specifying the external command to be used.}
}
\value{
The path to the created archive (invisibly).
}
\description{
A wrapper for an external \code{zip} command to create zip archives. Unlike
\code{\link[utils:zip]{utils::zip()}}, filepaths can be passed into the first argument with \verb{\%>\%}.
}
\examples{
\dontshow{.old_wd <- setwd(tempdir())}
# Zip a file to same basename
tmp <- path_temp("iris.csv")
write.csv(iris, tmp)
(zip <- zip_create(tmp))
file_size(c(tmp, zip))
\dontshow{setwd(.old_wd)}
}
